Implemented profile picture upload for organizations.

+++ lib/Organization/organization_dashboard.dart
@@ -52,8 +52,8 @@ class _OrganizationDashboardState extends State<OrganizationDashboard> {
       Campaign campaign = Campaign(
           title: element.data()['title'],
           description: element.data()['description'],
-          goalAmount: element.data()['goalAmount'],
-          amountRaised: element.data()['amountRaised'],
+          goalAmount: element.data()['goalAmount'].toDouble(),
+          amountRaised: element.data()['amountRaised'].toDouble(),
           category: element.data()['category'],
           endDate: element.data()['endDate'],
           dateCreated: element.data()['dateCreated'],
@@ -75,8 +75,8 @@ class _OrganizationDashboardState extends State<OrganizationDashboard> {
       UrgentCase urgentCase = UrgentCase(
           title: element.data()['title'],
           description: element.data()['description'],
-          goalAmount: element.data()['goalAmount'],
-          amountRaised: element.data()['amountRaised'],
+          goalAmount: element.data()['goalAmount'].toDouble(),
+          amountRaised: element.data()['amountRaised'].toDouble(),
           category: element.data()['category'],
           endDate: element.data()['endDate'],
           dateCreated: element.data()['dateCreated'],
@@ -98,8 +98,8 @@ class _OrganizationDashboardState extends State<OrganizationDashboard> {
       Beneficiary beneficiary = Beneficiary(
           name: element.data()['name'],
           biography: element.data()['biography'],
-          goalAmount: element.data()['goalAmount'],
-          amountRaised: element.data()['amountRaised'],
+          goalAmount: element.data()['goalAmount'].toDouble(),
+          amountRaised: element.data()['amountRaised'].toDouble(),
           category: element.data()['category'],
           endDate: element.data()['endDate'],
           dateCreated: element.data()['dateCreated'],
